+++ simpletr64/actions/wifi.py
@@ -13,12 +13,12 @@ class Wifi(DeviceTR64):
     """Class to get Wifi information's of a device which supports ``urn:dslforum-org:service:WLANConfiguration:X``.
 
     The class supports devices which supports ``urn:dslforum-org:service:WLANConfiguration:X`` namespace. Unless the
-    device is a AVM Fritz Box the DeviceTR64 objects needs to load the device definitions with
-    :meth:`~simpletr64.DeviceTR64.loadDeviceDefinitions` before the usage of any of the methods. For a Fritz.box
-    :meth:`~simpletr64.DeviceTR64.setupTR64Device` has to be called. Also a device might
-    not support all of the actions. This class does not implement all of the actions of this namespace, please
-    check the SCPD definitions if you miss some functionality. This library provides some tools to gather the
-    needed information's.
+    device is a AVM Fritz product the object needs to load the device definitions with
+    :meth:`~simpletr64.DeviceTR64.loadDeviceDefinitions` before the usage of any of the methods.
+    For a Fritz product :meth:`~simpletr64.DeviceTR64.setupTR64Device` can be called also this might not be future
+    compatible. Also a device might not support all of the actions. This class does not implement all of the actions
+    of this namespace, please check the SCPD definitions if you miss some functionality. This library provides some
+    tools to gather the needed information's.
 
     All Wifi actions ask for a interface id, this depends on the device if the counting starts with 0 or 1.
     Often a device supports more than one interface as for example to support 2.4 and 5 Ghz.